+++ benchmarks/benchmarks/ci/esmf_regridder.py
@@ -1,0 +1,39 @@
+"""Quick running benchmarks for :mod:`esmf_regrid.esmf_regridder`."""
+
+import numpy as np
+
+from esmf_regrid.esmf_regridder import GridInfo
+
+
+def _make_small_grid_args():
+    """
+    Not importing the one in test_GridInfo - if that changes, these benchmarks
+    would 'invisibly' change too.
+
+    """
+    small_x = 2
+    small_y = 3
+    small_grid_lon = np.array(range(small_x)) / (small_x + 1)
+    small_grid_lat = np.array(range(small_y)) * 2 / (small_y + 1)
+
+    small_grid_lon_bounds = np.array(range(small_x + 1)) / (small_x + 1)
+    small_grid_lat_bounds = np.array(range(small_y + 1)) * 2 / (small_y + 1)
+    return (
+        small_grid_lon,
+        small_grid_lat,
+        small_grid_lon_bounds,
+        small_grid_lat_bounds,
+    )
+
+
+class TimeGridInfo:
+    def setup(self):
+        lon, lat, lon_bounds, lat_bounds = _make_small_grid_args()
+        self.grid = GridInfo(lon, lat, lon_bounds, lat_bounds)
+
+    def time_make_grid(self):
+        """Basic test for :meth:`~esmf_regrid.esmf_regridder.GridInfo.make_esmf_field`."""
+        esmf_grid = self.grid.make_esmf_field()
+        esmf_grid.data[:] = 0
+
+    time_make_grid.version = 1
